+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+<!-- TODO -->
 - Added `skip_fileserver_update` setting (enabled by default) to speed up highstate rendering for a large number of minions. The setting can be disabled in `slskit.yaml` config:
 
 ```yaml